+++ ichnaea/models/constants.py
@@ -69,7 +69,8 @@ CELL_MAX_RADIUS = 100000  #: Max radius of a single cell network.
 WIFI_MAX_RADIUS = 5000  #: Max radius of a single WiFi network.
 
 MAX_OBSERVATION_AGE = 20000.0  #: Maximum observation age.
-MAX_OBSERVATION_ACCURACY = 200.0  #: Maximum observation age.
+MAX_OBSERVATION_ACCURACY = 200.0  #: Maximum observation accuracy.
+MAX_OBSERVATION_SPEED = 50.0  #: Maximum observation speed.
 
 WIFI_TEST_MAC = '01005e901000'
 """